+++ employee/views.py
@@ -1,3 +1,15 @@
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 
-# Create your views here.
+from .models import Employee
+
+
+def employee_index_view(request):
+    employee_list = Employee.objects.all()
+    return render(request, 'employee/index.html',
+                  {'employee_list': employee_list})
+
+
+def employee_detail_view(request, employee_id):
+    employee = get_object_or_404(Employee, pk=employee_id)
+    return render(request, 'employee/detail.html', {'employee': employee})
+
